+++ securedrop/source_app/views.py
@@ -1,7 +1,14 @@
-from flask import (Blueprint, render_template, flash, redirect, url_for,
-                   session)
-from flask_babel import gettext
+import os
 
+from flask import (Blueprint, render_template, flash, redirect, url_for,
+                   session, current_app)
+from flask_babel import gettext
+from sqlalchemy.exc import IntegrityError
+
+import crypto_util
+import store
+
+from db import Source, db_session
 from source_app.utils import logged_in, generate_unique_codename
 
 
@@ -29,5 +36,23 @@ def _main_blueprint():
         codename = generate_unique_codename()
         session['codename'] = codename
         return render_template('generate.html', codename=codename)
+
+    @view.route('/create', methods=['POST'])
+    def create():
+        filesystem_id = crypto_util.hash_codename(session['codename'])
+
+        source = Source(filesystem_id, crypto_util.display_id())
+        db_session.add(source)
+        try:
+            db_session.commit()
+        except IntegrityError as e:
+            current_app.logger.error(
+                "Attempt to create a source with duplicate codename: %s" %
+                (e,))
+        else:
+            os.mkdir(store.path(filesystem_id))
+
+        session['logged_in'] = True
+        return redirect(url_for('lookup'))
 
     return view
